+++ docker-compose-dev.yaml
@@ -39,7 +39,7 @@ services:
         volumes:
             - ./postgres:/docker-entrypoint-initdb.d
         ports:
-            - 5432:5432
+            - 5432
         networks:
             - basic_net
     adminer:
