+++ gui/src/dashboard_page/tooling/tooling.module.css
@@ -1,0 +1,30 @@
+.dropdown {
+    transition: all 0.5s ease-in-out;
+    position: absolute;
+    top: -5px;
+    left: 30%;
+    z-index: 5;
+    border-radius: 10px;
+    background-color: rgb(24, 24, 27);
+    width: 40%;
+}
+.dropdownContent {
+    min-height: 0px;
+    height: 0px;
+    transition: all 0.5s ease-in-out;
+    overflow: hidden;
+}
+
+.dropdown:hover .dropdownContent {
+    height: 600px;
+}
+
+.toolingBoundingBox {
+    overflow: auto;
+    height: 500px;
+}
+
+.tooling {
+    width: 100%;
+    height: "100%";
+}
